feat(smart-apply): Add selection prompt for the custom model
